Add custom type for tracked certificates
I wrote custom type and provider for tracked certificates which
is able to display in puppet tracked certificates that have been
both rejected and accepted. On the other hand, the provider is
also able to request new certificates and stop tracking certificates
if the :enabled flag is appropriate.

Note that this still doesn't handle files existing already nor does
it have the option (yet) to fail if the CA has rejected the
certificate.